修正

+++ app/controllers/items_controller.rb
@@ -35,9 +35,11 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to action: :index
+    if current_user.id == @item.user_id
+      item = Item.find(params[:id])
+      item.destroy
+      redirect_to action: :index
+    end 
   end
 
   private
